chore: 他のa11y系ruleで利用できるようにstyle-componentsの拡張などに制限を入れる仕組みを切り出す
